LayoutTests/fast/exclusions/css-exclusions-enabled-expected.txt

Test that parsing of css exclusions related properties is enabled by default.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS testInlineStyle("-webkit-wrap-flow", "both") is "both"
PASS testComputedStyle("-webkit-wrap-flow", "both") is "both"
PASS testInlineStyle("-webkit-wrap-through", "none") is "none"
PASS testComputedStyle("-webkit-wrap-through", "none") is "none"
PASS successfullyParsed is true

TEST COMPLETE
